feat(db)!: migrate the database engine from PGlite to SQLite (node:sqlite)
Replace the PGlite WASM Postgres engine with SQLite via Electron's built-in
node:sqlite (DatabaseSync), keeping drizzle-orm as the query layer. Decision
and evidence: sqlite-poc/RESULTS.md (PGlite->SQLite POC with adversarial
benchmarks); full audit trail: sqlite-migration-progress.md.

WHY (measured on the real workload, 50k-song library):
- Engine open: 424-428ms (PGlite, clean) vs 4ms (SQLite); PGlite was
  24-28s on churned production DBs.
- Idle main-process commit: -106MB vs PGlite; removes the ~194MB
  non-collectable postgres.wasm heap entirely.
- Fuzzy search: 489-1279ms -> 0.06-13ms per query.
- Nora-shaped ingestion (batch tx + per-track savepoints): 234 -> 4426 tracks/s.
- DB size @50k: 385MB (PGlite dir) -> 47MB (single SQLite file).
- Savepoint/transaction semantics, crash recovery, and forensic invariants
  all preserved (gates below).

ENGINE (src/main/db/sqlite/):
- engine.ts: node:sqlite DatabaseSync + drizzle sqlite-proxy adapter.
  PRAGMAs: WAL, synchronous=FULL (power-loss durability; measured cost
  acceptable for Nora's batch-commit pattern), busy_timeout=5s,
  foreign_keys=ON, temp_store=MEMORY, 16MB page cache. Baseline DDL applied
  once, stamped via PRAGMA user_version.
  Transaction serialization reproduces PGlite's contract: drizzle's
  sqlite-proxy does NOT serialize transactions (overlapping BEGINs collide,
  bare statements can join an open transaction), so orm.transaction is
  FIFO-queued holding the lock across the whole async body, and bare
  statements wait at execution time while a transaction is active. Nested
  savepoints run on the unpatched trx inside the held lock (no deadlock).
- ddl.ts: baseline schema for all 42 tables. *_ci columns are
  COLLATE NOCASE GENERATED ALWAYS AS (lower(x)) STORED (citext parity);
  *_norm columns strip punctuation/space (mirrors pg regexp_replace +
  space removal) for punctuation-insensitive search. FTS5 trigram tables
  (songs/albums/artists/genres/playlists, raw + norm) kept in sync by
  triggers; AFTER INSERT trigger assigns metadata_undo_snapshots.seq
  (MAX+1) while preserving explicit migrated values.
- raw.ts: object-row raw queries (drizzle's sqlite-proxy returns positional
  arrays for raw .all(); helpers target the same connection so they are
  transaction-safe).
- migrate-from-pglite.ts: one-shot legacy-data migration. Runs crash-atomically
  in a single transaction (any crash rolls back everything; the empty-DB
  residue check re-migrates on next boot). FKs disabled during bulk insert
  (out-of-order self-referencing music_folders), re-enabled + PRAGMA
  foreign_key_check afterwards. Residue detection includes user_settings
  (seeded only after a completed migration+boot) so an empty-library user
  does not resurrect their old library. Timestamps read as ::text and parsed
  as UTC (PGlite raw parse applies local offset; drizzle-pg applies UTC);
  ids and undo-snapshot seq values preserved; rows keyed by drizzle property
  names; generated columns recomputed by SQLite.

SCHEMA (src/main/db/schema.ts): full rewrite to drizzle sqlite-core — 42
tables, all relations ported 1:1, identical exported symbols. Mapping:
identity PK -> INTEGER PRIMARY KEY (rowid), citext -> NOCASE lower()
generated columns, numeric -> REAL, timestamps -> epoch-ms
(timestamp_ms), booleans -> INTEGER mode, pgEnum -> TEXT + CHECK (in DDL),
json/jsonb -> TEXT json-mode. Second identity column
(metadata_undo_snapshots.seq) replaced by trigger assignment.

QUERY LAYER: all PostgreSQL dialect constructs rewritten — ::casts removed,
to_char/extract -> strftime, NOW()/INTERVAL -> compile-time ms cutoffs
(removes pg session-TZ dependence), UPDATE..FROM (VALUES) -> CTE form,
btrim -> trim, pg affectedRows -> sqlite changes, overridingSystemValue
dropped (SQLite rowid accepts explicit ids), db.execute -> raw helpers.
Latent bug fixed: artists.ts raw SQL referenced non-existent "artistsSongs"
table.

SEARCH (src/main/search/): five engines rewritten — raw LIKE + *_norm LIKE
for exact/prefix/substring/punctuation-insensitive matching; pg_trgm %
fuzzy branch replaced by FTS5 trigram-OR candidate pools scored in JS with
pg_trgm's exact algorithm (search/fuzzy/pgTrgmSimilarity.ts, verified to
reproduce PG's match sets exactly); computeTier/normalizeQuery unchanged.
POC-verified recall 1.00 vs pg_trgm on every probe type, 5-100x faster.
SmartPlaylistCompiler: ILIKE -> lower() LIKE, interval cutoffs -> epoch-ms.

SEARCH CONTRACT NOTE: SQLite lower()/NOCASE are ASCII-only (pg citext folds
Unicode). ASCII CI parity holds; non-ASCII handled by *_norm columns +
JS normalizeForSearch. Duration columns are REAL (numbers) where pg numeric
returned strings — call sites coerced.

EXPORT/IMPORT: pgDump replaced by a portable SQL dump (DELETE+INSERT replace
semantics, generated columns excluded) + importDatabase(). Export files
keep the same user flow (import accepts the legacy filename too).

LIFECYCLE: db.ts keeps its exports (db, DB_PATH, closeDatabaseInstance,
isDatabaseStubbed, nukeDatabase, exportDatabase, importDatabase). A legacy
nora.pglite.db/ dir is migrated automatically on first launch, then kept as
backup. NORA_USE_PGLITE=1 now throws (rollback path = master branch).
NORA_DB_FILE env lets integration tests drive the production file path.

TESTS:
- test/helpers/sqliteMockDb.ts + @test-helpers alias: in-memory SQLite mock
  replacing per-file PGlite factories; 12+ test files ported.
- New gates: test/src/main/db/sqlite/sqlite-engine.test.ts (DDL↔drizzle
  parity 42 tables/334 columns, generated columns, FTS sync, seq trigger,
  export/import round-trip, REAL PGlite->SQLite migration fidelity) and
  smoke.test.ts (production-path ingest -> search -> favorite -> playlist ->
  smart rules -> restart persistence, no db mocks).
- P1 savepoint atomicity 4/4, P0 mass-deletion, backpressure fallback,
  adversarial shutdown, GateD4 concurrency/rollback — all PASS on SQLite.
- Full suite: 353+/358 files, 2221+/2235 tests pass; the 5 remaining failing
  files (miniPlayerGeometry, songWorkerPool, tryToParseSong, waveformJob,
  autotagArtwork) fail identically on pristine master (pre-existing).

BUILD/PACKAGE: electron-vite + electron-builder --dir clean; node:sqlite
verified inside the packaged Nora.exe binary and an asar-packaged app main
process; packaged launch smoke booted the real app on SQLite (schema 669ms,
seeded, integrity ok).

KNOWN PRE-EXISTING (NOT from this migration): miniPlayerGeometry,
songWorkerPool, tryToParseSong, waveformJob, autotagArtwork test failures
and renderer typecheck errors — verified on the pristine branch.

Also includes sqlite-poc/ (the benchmark POC harness and reports that
justified this decision) and sqlite-migration-progress.md (audit journal).

+++ src/main/db/queries/history.ts
@@ -55,7 +55,7 @@ export const getAllSongsInHistory = async (
       const historyRecords = await trx
         .select({
           songId: playHistory.songId,
-          playCount: sql<number>`count(*)::int`,
+          playCount: sql<number>`count(*)`,
           lastPlayed: sql<Date>`max(${playHistory.createdAt})`
         })
         .from(playHistory)
